EMS 24 Modelo Nota Fiscal

+++ nds-client/src/test/java/br/com/abril/nds/fixture/Fixture.java
@@ -2964,7 +2964,7 @@ public class Fixture {
 			Long ncm,
 			NotaFiscal notaFiscal,
 			ProdutoEdicao produtoEdicao,
-			Long quantidade,
+			BigDecimal quantidade,
 			String unidade,
 			BigDecimal valorDesconto,
 			BigDecimal valorFrete,
